+++ drake/systems/framework/continuous_system.h
@@ -34,7 +34,7 @@ class ContinuousSystem : public ContinuousSystemInterface<T> {
           "MapVelocityToConfigurationDerivatives?");
     }
 
-    for (int64_t i = 0; i < generalized_velocity.size(); ++i) {
+    for (int i = 0; i < generalized_velocity.size(); ++i) {
       configuration_derivatives->SetAtIndex(i,
                                             generalized_velocity.GetAtIndex(i));
     }
