+++ drake/systems/framework/test/basic_state_vector_test.cc
@@ -9,7 +9,7 @@ namespace drake {
 namespace systems {
 namespace {
 
-const int64_t kLength = 2;
+const int kLength = 2;
 
 class BasicStateVectorTest : public ::testing::Test {
  protected:
